不能写入TF卡 TF读写测试没问题，连续两次也可以 之后能读卡不能写卡 写卡出现FR_INVALID_OBJECT(9),随即单片机跑飞。

+++ src/includes.h
@@ -23,7 +23,8 @@ typedef signed short SWORD;
 
 #ifndef _TYPEDEF_DWORD_
 #define _TYPEDEF_DWORD_
-typedef unsigned int DWORD;
+//typedef unsigned int DWORD;
+typedef unsigned long int DWORD;
 #endif
 
 #if 0
